refactor: split db schema into modules
